requirements.txt
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
anthropic
tqdm
rank_bm25